#step1 [refactor] split 테스트를 Input을 사용해서 하도록 수정
기존에 StringCalculator.split을 테스트 했었지만, 이제 이 부분은 Input 객체가 맡게 되어 테스트 위치를 옮기고 Input.split을 사용하도록 수정
