+++ src/test/java/calculator/InputTest.java
@@ -1,11 +1,19 @@
 package calculator;
 
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Java6Assertions.assertThat;
 
 public class InputTest {
+
+    private String[] result;
+
+    @BeforeEach
+    void setUp() {
+        result = new String[]{"1", "2"};
+    }
 
     @Test
     void isEmpty() {
@@ -14,5 +22,26 @@ public class InputTest {
 
         Input input2 = new Input(null);
         assertThat(input2.isEmpty()).isTrue();
+    }
+
+    @Test
+    @DisplayName("사용자 입력값 split (with comma)")
+    void split() {
+        Input input = new Input("1,2");
+        assertThat(input.split()).containsExactly(result);
+    }
+
+    @Test
+    @DisplayName("사용자 입력값 split (with colon)")
+    void split2() {
+        Input input = new Input("1:2");
+        assertThat(input.split()).containsExactly(result);
+    }
+
+    @Test
+    @DisplayName("사용자 입력값 split (with custom delimiter")
+    void split3() {
+        Input input = new Input("//;\n1;2");
+        assertThat(input.split()).containsExactly(result);
     }
 }
